+++ dependencies/src/main/java/com/google/cloud/tools/opensource/classpath/ClassDumper.java
@@ -92,14 +92,10 @@ class ClassDumper {
     
     Map<String, ClassPathEntry> map = new HashMap<>();
     for (ClassPathEntry entry : entries) {
-      try {
-        for (String className : entry.getFileNames()) {
-          if (!map.containsKey(className)) {
-            map.put(className, entry);
-          }
+      for (String className : entry.getFileNames()) {
+        if (!map.containsKey(className)) {
+          map.put(className, entry);
         }
-      } catch (IOException ex) {
-        throw new IOException("Failed to read content of " + entry.getJar(), ex);
       }
     }
